+++ OrionBE.Launcher/Services/IStartupDependencyCheckService.cs
@@ -1,0 +1,6 @@
+namespace OrionBE.Launcher.Services;
+
+public interface IStartupDependencyCheckService
+{
+    Task<StartupDependencyCheckReport> RunIfFirstLaunchAsync(CancellationToken cancellationToken = default);
+}